modify query11 cypher impl to add third sort order, according to spec modify query12 cypher impl to be consistent with spec update tests accordingly

+++ src/main/java/com/ldbc/socialnet/workload/neo4j/interactive/embedded_cypher/Neo4jQuery12EmbeddedCypher.java
@@ -33,7 +33,7 @@ public class Neo4jQuery12EmbeddedCypher extends Neo4jQuery12<ExecutionEngine> {
                                 (String) row.get("friendFirstName"),
                                 (String) row.get("friendLastName"),
                                 (Collection<String>) row.get("tagNames"),
-                                (int) row.get("count"));
+                                ((Long) row.get("count")).intValue());
                     }
                 });
     }
